app/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
werkzeug
Pillow

python-magic
pydantic
pydantic-settings
python-multipart
fastapi-limiter
redis
ruff
jinja2
celery